Fixed/Added Lots.
Fixed both food, and material spawnning, aswell as platform collision, and platform size.

Changed the way the flood works, and the way energy is lost. Now you increase the flood each time you hit the floor, and you run away from it. The flood also does not continue to move more than 100 pixels below the screen.

Modified the game so that you are always jumping, and only have to move left to right